Updated dockerfile
