+++ ndb_api/interfaces.go
@@ -18,18 +18,12 @@ package ndb_api
 
 import (
 	"context"
-
-	"github.com/nutanix-cloud-native/ndb-operator/api/v1alpha1"
 )
 
 type ProfileResolver interface {
 	Resolve(ctx context.Context, allProfiles []ProfileResponse, filter func(p ProfileResponse) bool) (profile ProfileResponse, err error)
 	GetName() string
 	GetId() string
-}
-
-type DatabaseActionArgs interface {
-	Get(dbSpec v1alpha1.DatabaseSpec) []ActionArgument
 }
 
 type ProfileResolvers map[string]ProfileResolver
