+++ wrapper_implementations/database.go
@@ -1,7 +1,22 @@
-package controller_types
+/*
+Copyright 2022-2023 Nutanix, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package wrapper_implementations
 
 import (
-	"errors"
 	"strings"
 
 	"github.com/nutanix-cloud-native/ndb-operator/api/v1alpha1"
@@ -9,6 +24,9 @@ import (
 	"github.com/nutanix-cloud-native/ndb-operator/ndb_api"
 )
 
+// Wrapper over api/v1alpha1.Database
+// required to provide implementation of the
+// DatabaseInterface defined in the package ndb_api
 type Database struct {
 	v1alpha1.Database
 }
@@ -39,6 +57,7 @@ func (d *Database) GetNDBClusterId() string {
 
 func (d *Database) GetDBInstanceActionArguments() []ndb_api.ActionArgument {
 	dbTypeActionArgs, _ := GetActionArgumentsByDatabaseType(d.GetDBInstanceType())
+	// TODO: Handle error
 	return dbTypeActionArgs.Get(d.Spec)
 }
 
@@ -64,20 +83,4 @@ func (d *Database) GetProfileResolvers() ndb_api.ProfileResolvers {
 
 	return profileResolvers
 
-}
-
-// Returns action arguments based on the type of database
-func GetActionArgumentsByDatabaseType(databaseType string) (ndb_api.DatabaseActionArgs, error) {
-	var dbTypeActionArgs ndb_api.DatabaseActionArgs
-	switch databaseType {
-	case common.DATABASE_TYPE_MYSQL:
-		dbTypeActionArgs = &MysqlActionArgs{}
-	case common.DATABASE_TYPE_POSTGRES:
-		dbTypeActionArgs = &PostgresActionArgs{}
-	case common.DATABASE_TYPE_MONGODB:
-		dbTypeActionArgs = &MongodbActionArgs{}
-	default:
-		return nil, errors.New("invalid database type: supported values: mysql, postgres, mongodb")
-	}
-	return dbTypeActionArgs, nil
 }
